rumi

+++ dev/application/views/Admin/managePage.php
@@ -56,10 +56,10 @@
                                 <td><?php echo $pd->insertedBy?></td>
                                 <td><?php if ($pd->lastModifiedBy==""){echo NEVER_MODIFIED;}else{echo $pd->lastModifiedBy;} ?></td>
                                 <td><?php if ($pd->lastModifiedDate==""){echo NEVER_MODIFIED;}
-                                    else{
-                                        $timestamp = strtotime($pd->lastModifiedDate);
-                                        $date = date('d-m-Y', $timestamp);
-                                        echo $date ;}
+                                    else
+                                    {
+                                        echo preg_replace("/ /","<br>",date('d-m-Y h:i A',strtotime($pd->lastModifiedDate)),1);
+                                    }
                                     ?>
                                     </td>
                                 <td><?php echo $pd->pageStatus?></td>
@@ -67,7 +67,7 @@
                                     <div class="btn-group">
 
                                         <a class="btn" href="<?php echo base_url()?>Admin/Page/editPageShow/<?php echo $pd->pageId?>"><i class="icon_pencil-edit"></i></a>
-                                        <a class="btn " data-panel-id="<?php echo $pd->pageId ?>"  onclick="" href="<?php echo base_url()?>Admin/Page/deletePage/<?php echo $pd->pageId?>"><i class="icon_trash"></i></a>
+                                        <a class="btn " data-panel-id="<?php echo $pd->pageId ?>"  onclick='return confirm("Are you sure?")' href="<?php echo base_url()?>Admin/Page/deletePage/<?php echo $pd->pageId?>"><i class="icon_trash"></i></a>
 
                                     </div>
                                 </td>
